requirements.txt
pydantic>=2.0
PyYAML>=6.0
gitpython>=3.1.0
rich>=13.0
pytest>=7.0
google-generativeai>=0.3.0
